+++ backend/infrahub/core/diff/query/diff_get.py
@@ -1,27 +1,11 @@
-from typing import Any, Iterable
+from typing import Any
 
-from neo4j.graph import Node as Neo4jNode
-from neo4j.graph import Path as Neo4jPath
-
-from infrahub.core.constants import DiffAction
-from infrahub.core.constants.database import DatabaseEdgeType
-from infrahub.core.query import Query, QueryResult, QueryType
+from infrahub.core.query import Query, QueryType
 from infrahub.core.timestamp import Timestamp
 from infrahub.database import InfrahubDatabase
 
-from ..model.path import (
-    BranchTrackingId,
-    ConflictSelection,
-    EnrichedDiffAttribute,
-    EnrichedDiffConflict,
-    EnrichedDiffNode,
-    EnrichedDiffProperty,
-    EnrichedDiffRelationship,
-    EnrichedDiffRoot,
-    EnrichedDiffSingleRelationship,
-    NameTrackingId,
-    TrackingId,
-)
+from ..model.path import EnrichedDiffRoot, TrackingId
+from ..repository.deserializer import EnrichedDiffDeserializer
 from .filters import EnrichedDiffQueryFilters
 
 QUERY_MATCH_NODES = """
@@ -59,6 +43,7 @@ class EnrichedDiffGetQuery(Query):
 
     def __init__(
         self,
+        deserializer: EnrichedDiffDeserializer,
         base_branch_name: str,
         diff_branch_names: list[str],
         filters: EnrichedDiffQueryFilters,
@@ -69,6 +54,7 @@ class EnrichedDiffGetQuery(Query):
         **kwargs: Any,
     ) -> None:
         super().__init__(**kwargs)
+        self.deserializer = deserializer
         self.base_branch_name = base_branch_name
         self.diff_branch_names = diff_branch_names
         self.from_time: Timestamp = from_time or Timestamp("2000-01-01T00:00:01Z")
@@ -192,334 +178,7 @@ class EnrichedDiffGetQuery(Query):
         self.order_by = ["diff_root.diff_branch_name ASC", "diff_root.from_time ASC", "diff_node.label ASC"]
 
     async def get_enriched_diff_roots(self, include_parents: bool) -> list[EnrichedDiffRoot]:
-        deserializer = EnrichedDiffDeserializer()
-        enriched_diffs = await deserializer.deserialize(
+        enriched_diffs = await self.deserializer.deserialize(
             database_results=self.get_results(), include_parents=include_parents
         )
         return enriched_diffs
-
-
-class EnrichedDiffDeserializer:
-    def __init__(self) -> None:
-        self._diff_root_map: dict[str, EnrichedDiffRoot] = {}
-        self._diff_node_map: dict[tuple[str, str], EnrichedDiffNode] = {}
-        self._diff_node_attr_map: dict[tuple[str, str, str], EnrichedDiffAttribute] = {}
-        self._diff_node_rel_group_map: dict[tuple[str, str, str], EnrichedDiffRelationship] = {}
-        self._diff_node_rel_element_map: dict[tuple[str, str, str, str], EnrichedDiffSingleRelationship] = {}
-        self._diff_prop_map: dict[tuple[str, str, str, str] | tuple[str, str, str, str, str], EnrichedDiffProperty] = {}
-
-    async def deserialize(
-        self, database_results: Iterable[QueryResult], include_parents: bool
-    ) -> list[EnrichedDiffRoot]:
-        results = list(database_results)
-        for result in results:
-            enriched_root = self._deserialize_diff_root(root_node=result.get_node("diff_root"))
-            node_node = result.get(label="diff_node")
-            if not isinstance(node_node, Neo4jNode):
-                continue
-            enriched_node = self._deserialize_diff_node(node_node=node_node, enriched_root=enriched_root)
-            node_conflict_node = result.get(label="diff_node_conflict")
-            if isinstance(node_conflict_node, Neo4jNode) and not enriched_node.conflict:
-                self._deserialize_conflict(diff_conflict_node=node_conflict_node, linked_node=enriched_node)
-            self._deserialize_attributes(result=result, enriched_root=enriched_root, enriched_node=enriched_node)
-            self._deserialize_relationships(result=result, enriched_root=enriched_root, enriched_node=enriched_node)
-
-        if include_parents:
-            for result in results:
-                enriched_root = self._deserialize_diff_root(root_node=result.get_node("diff_root"))
-                self._deserialize_parents(result=result, enriched_root=enriched_root)
-
-        return list(self._diff_root_map.values())
-
-    def _deserialize_attributes(
-        self, result: QueryResult, enriched_root: EnrichedDiffRoot, enriched_node: EnrichedDiffNode
-    ) -> None:
-        for attribute_result in result.get_nested_node_collection("diff_attributes"):
-            diff_attr_node, diff_attr_property_node, diff_attr_property_conflict = attribute_result
-            if diff_attr_node is None or diff_attr_property_node is None:
-                continue
-            enriched_attribute = self._deserialize_diff_attr(
-                diff_attr_node=diff_attr_node, enriched_root=enriched_root, enriched_node=enriched_node
-            )
-            enriched_property = self._deserialize_diff_attr_property(
-                diff_attr_property_node=diff_attr_property_node,
-                enriched_attr=enriched_attribute,
-                enriched_node=enriched_node,
-                enriched_root=enriched_root,
-            )
-            if diff_attr_property_conflict:
-                self._deserialize_conflict(
-                    diff_conflict_node=diff_attr_property_conflict, linked_node=enriched_property
-                )
-
-    def _deserialize_relationships(
-        self, result: QueryResult, enriched_root: EnrichedDiffRoot, enriched_node: EnrichedDiffNode
-    ) -> None:
-        for relationship_result in result.get_nested_node_collection("diff_relationships"):
-            group_node, element_node, element_conflict, property_node, property_conflict = relationship_result
-            if group_node is not None and element_node is None and property_node is None:
-                enriched_node.add_relationship_from_DiffRelationship(diff_rel=group_node)
-                continue
-            if group_node is None or element_node is None or property_node is None:
-                continue
-
-            enriched_relationship_group = self._deserialize_diff_relationship_group(
-                relationship_group_node=group_node, enriched_root=enriched_root, enriched_node=enriched_node
-            )
-            enriched_relationship_element = self._deserialize_diff_relationship_element(
-                relationship_element_node=element_node,
-                enriched_relationship_group=enriched_relationship_group,
-                enriched_node=enriched_node,
-                enriched_root=enriched_root,
-            )
-            if element_conflict and not enriched_relationship_element.conflict:
-                self._deserialize_conflict(
-                    diff_conflict_node=element_conflict, linked_node=enriched_relationship_element
-                )
-            element_property = self._deserialize_diff_relationship_element_property(
-                relationship_element_property_node=property_node,
-                enriched_relationship_element=enriched_relationship_element,
-                enriched_relationship_group=enriched_relationship_group,
-                enriched_node=enriched_node,
-                enriched_root=enriched_root,
-            )
-            if property_conflict:
-                self._deserialize_conflict(diff_conflict_node=property_conflict, linked_node=element_property)
-
-    def _deserialize_parents(self, result: QueryResult, enriched_root: EnrichedDiffRoot) -> None:
-        parents_path = result.get("parents_path")
-        if not parents_path or not isinstance(parents_path, Neo4jPath):
-            return
-
-        node_uuid = result.get(label="diff_node").get("uuid")
-
-        # Remove the node itself from the path
-        parents_path = parents_path.nodes[1:]  # type: ignore[union-attr]
-
-        # TODO Ensure the list is even
-        current_node_uuid = node_uuid
-        for rel, parent in zip(parents_path[::2], parents_path[1::2]):
-            enriched_root.add_parent(
-                node_id=current_node_uuid,
-                parent_id=parent.get("uuid"),
-                parent_kind=parent.get("kind"),
-                parent_label=parent.get("label"),
-                parent_rel_name=rel.get("name"),
-                parent_rel_label=rel.get("label"),
-            )
-            current_node_uuid = parent.get("uuid")
-
-    def _get_str_or_none_property_value(self, node: Neo4jNode, property_name: str) -> str | None:
-        value_raw = node.get(property_name)
-        return str(value_raw) if value_raw is not None else None
-
-    def _deserialize_diff_root(self, root_node: Neo4jNode) -> EnrichedDiffRoot:
-        root_uuid = str(root_node.get("uuid"))
-        if root_uuid in self._diff_root_map:
-            return self._diff_root_map[root_uuid]
-
-        from_time = Timestamp(str(root_node.get("from_time")))
-        to_time = Timestamp(str(root_node.get("to_time")))
-        tracking_id_str = self._get_str_or_none_property_value(node=root_node, property_name="tracking_id")
-        tracking_id = None
-        if tracking_id_str:
-            for tracking_id_class in (BranchTrackingId, NameTrackingId):
-                try:
-                    tracking_id = tracking_id_class.deserialize(id_string=tracking_id_str)
-                    break
-                except ValueError:
-                    ...
-        enriched_root = EnrichedDiffRoot(
-            base_branch_name=str(root_node.get("base_branch")),
-            diff_branch_name=str(root_node.get("diff_branch")),
-            from_time=from_time,
-            to_time=to_time,
-            uuid=str(root_uuid),
-            tracking_id=tracking_id,
-            num_added=int(root_node.get("num_added")),
-            num_updated=int(root_node.get("num_updated")),
-            num_removed=int(root_node.get("num_removed")),
-            num_conflicts=int(root_node.get("num_conflicts")),
-            contains_conflict=str(root_node.get("contains_conflict")).lower() == "true",
-        )
-        self._diff_root_map[root_uuid] = enriched_root
-        return enriched_root
-
-    def _deserialize_diff_node(self, node_node: Neo4jNode, enriched_root: EnrichedDiffRoot) -> EnrichedDiffNode:
-        node_uuid = str(node_node.get("uuid"))
-        node_key = (enriched_root.uuid, node_uuid)
-        if node_key in self._diff_node_map:
-            return self._diff_node_map[node_key]
-
-        timestamp_str = self._get_str_or_none_property_value(node=node_node, property_name="changed_at")
-        enriched_node = EnrichedDiffNode(
-            uuid=node_uuid,
-            kind=str(node_node.get("kind")),
-            label=str(node_node.get("label")),
-            changed_at=Timestamp(timestamp_str) if timestamp_str else None,
-            action=DiffAction(str(node_node.get("action"))),
-            path_identifier=str(node_node.get("path_identifier")),
-            num_added=int(node_node.get("num_added")),
-            num_updated=int(node_node.get("num_updated")),
-            num_removed=int(node_node.get("num_removed")),
-            num_conflicts=int(node_node.get("num_conflicts")),
-            contains_conflict=str(node_node.get("contains_conflict")).lower() == "true",
-        )
-        self._diff_node_map[node_key] = enriched_node
-        enriched_root.nodes.add(enriched_node)
-        return enriched_node
-
-    def _deserialize_diff_attr(
-        self, diff_attr_node: Neo4jNode, enriched_root: EnrichedDiffRoot, enriched_node: EnrichedDiffNode
-    ) -> EnrichedDiffAttribute:
-        attr_name = str(diff_attr_node.get("name"))
-        attr_key = (enriched_root.uuid, enriched_node.uuid, attr_name)
-        if attr_key in self._diff_node_attr_map:
-            return self._diff_node_attr_map[attr_key]
-
-        enriched_attr = EnrichedDiffAttribute(
-            name=str(diff_attr_node.get("name")),
-            changed_at=Timestamp(str(diff_attr_node.get("changed_at"))),
-            path_identifier=str(diff_attr_node.get("path_identifier")),
-            action=DiffAction(str(diff_attr_node.get("action"))),
-            num_added=int(diff_attr_node.get("num_added")),
-            num_updated=int(diff_attr_node.get("num_updated")),
-            num_removed=int(diff_attr_node.get("num_removed")),
-            num_conflicts=int(diff_attr_node.get("num_conflicts")),
-            contains_conflict=str(diff_attr_node.get("contains_conflict")).lower() == "true",
-        )
-        self._diff_node_attr_map[attr_key] = enriched_attr
-        enriched_node.attributes.add(enriched_attr)
-        return enriched_attr
-
-    def _deserialize_diff_relationship_group(
-        self, relationship_group_node: Neo4jNode, enriched_root: EnrichedDiffRoot, enriched_node: EnrichedDiffNode
-    ) -> EnrichedDiffRelationship:
-        diff_rel_name = str(relationship_group_node.get("name"))
-        rel_key = (enriched_root.uuid, enriched_node.uuid, diff_rel_name)
-        if rel_key in self._diff_node_rel_group_map:
-            return self._diff_node_rel_group_map[rel_key]
-
-        enriched_relationship = EnrichedDiffRelationship.from_graph(node=relationship_group_node)
-        self._diff_node_rel_group_map[rel_key] = enriched_relationship
-        enriched_node.relationships.add(enriched_relationship)
-        return enriched_relationship
-
-    def _deserialize_diff_relationship_element(
-        self,
-        relationship_element_node: Neo4jNode,
-        enriched_relationship_group: EnrichedDiffRelationship,
-        enriched_node: EnrichedDiffNode,
-        enriched_root: EnrichedDiffRoot,
-    ) -> EnrichedDiffSingleRelationship:
-        diff_element_peer_id = str(relationship_element_node.get("peer_id"))
-        rel_element_key = (
-            enriched_root.uuid,
-            enriched_node.uuid,
-            enriched_relationship_group.name,
-            diff_element_peer_id,
-        )
-        if rel_element_key in self._diff_node_rel_element_map:
-            return self._diff_node_rel_element_map[rel_element_key]
-
-        peer_label = self._get_str_or_none_property_value(node=relationship_element_node, property_name="peer_label")
-        enriched_rel_element = EnrichedDiffSingleRelationship(
-            changed_at=Timestamp(str(relationship_element_node.get("changed_at"))),
-            action=DiffAction(str(relationship_element_node.get("action"))),
-            peer_id=diff_element_peer_id,
-            peer_label=peer_label,
-            path_identifier=str(relationship_element_node.get("path_identifier")),
-            num_added=int(relationship_element_node.get("num_added")),
-            num_updated=int(relationship_element_node.get("num_updated")),
-            num_removed=int(relationship_element_node.get("num_removed")),
-            num_conflicts=int(relationship_element_node.get("num_conflicts")),
-            contains_conflict=str(relationship_element_node.get("contains_conflict")).lower() == "true",
-        )
-        enriched_relationship_group.relationships.add(enriched_rel_element)
-        self._diff_node_rel_element_map[rel_element_key] = enriched_rel_element
-        return enriched_rel_element
-
-    def _property_node_to_enriched_property(self, property_node: Neo4jNode) -> EnrichedDiffProperty:
-        previous_value = self._get_str_or_none_property_value(node=property_node, property_name="previous_value")
-        new_value = self._get_str_or_none_property_value(node=property_node, property_name="new_value")
-        return EnrichedDiffProperty(
-            property_type=DatabaseEdgeType(str(property_node.get("property_type"))),
-            changed_at=Timestamp(str(property_node.get("changed_at"))),
-            previous_value=previous_value,
-            new_value=new_value,
-            action=DiffAction(str(property_node.get("action"))),
-            path_identifier=str(property_node.get("path_identifier")),
-        )
-
-    def _deserialize_diff_attr_property(
-        self,
-        diff_attr_property_node: Neo4jNode,
-        enriched_attr: EnrichedDiffAttribute,
-        enriched_node: EnrichedDiffNode,
-        enriched_root: EnrichedDiffRoot,
-    ) -> EnrichedDiffProperty:
-        diff_prop_type = str(diff_attr_property_node.get("property_type"))
-        attr_property_key = (enriched_root.uuid, enriched_node.uuid, enriched_attr.name, diff_prop_type)
-        if attr_property_key in self._diff_prop_map:
-            return self._diff_prop_map[attr_property_key]
-
-        enriched_property = self._property_node_to_enriched_property(property_node=diff_attr_property_node)
-        enriched_attr.properties.add(enriched_property)
-        self._diff_prop_map[attr_property_key] = enriched_property
-        return enriched_property
-
-    def _deserialize_diff_relationship_element_property(
-        self,
-        relationship_element_property_node: Neo4jNode,
-        enriched_relationship_element: EnrichedDiffSingleRelationship,
-        enriched_relationship_group: EnrichedDiffRelationship,
-        enriched_node: EnrichedDiffNode,
-        enriched_root: EnrichedDiffRoot,
-    ) -> EnrichedDiffProperty:
-        diff_prop_type = str(relationship_element_property_node.get("property_type"))
-        rel_property_key = (
-            enriched_root.uuid,
-            enriched_node.uuid,
-            enriched_relationship_group.name,
-            enriched_relationship_element.peer_id,
-            diff_prop_type,
-        )
-        if rel_property_key in self._diff_prop_map:
-            return self._diff_prop_map[rel_property_key]
-
-        enriched_property = self._property_node_to_enriched_property(property_node=relationship_element_property_node)
-        self._diff_prop_map[rel_property_key] = enriched_property
-        enriched_relationship_element.properties.add(enriched_property)
-        return enriched_property
-
-    def _deserialize_conflict(
-        self,
-        diff_conflict_node: Neo4jNode,
-        linked_node: EnrichedDiffNode | EnrichedDiffSingleRelationship | EnrichedDiffProperty,
-    ) -> EnrichedDiffConflict:
-        base_branch_value = self._get_str_or_none_property_value(
-            node=diff_conflict_node, property_name="base_branch_value"
-        )
-        diff_branch_value = self._get_str_or_none_property_value(
-            node=diff_conflict_node, property_name="diff_branch_value"
-        )
-        base_timestamp_str = self._get_str_or_none_property_value(
-            node=diff_conflict_node, property_name="base_branch_changed_at"
-        )
-        diff_timestamp_str = self._get_str_or_none_property_value(
-            node=diff_conflict_node, property_name="diff_branch_changed_at"
-        )
-        selected_branch = self._get_str_or_none_property_value(node=diff_conflict_node, property_name="selected_branch")
-        conflict = EnrichedDiffConflict(
-            uuid=str(diff_conflict_node.get("uuid")),
-            base_branch_action=DiffAction(str(diff_conflict_node.get("base_branch_action"))),
-            base_branch_value=base_branch_value,
-            base_branch_changed_at=Timestamp(base_timestamp_str) if base_timestamp_str else None,
-            diff_branch_action=DiffAction(str(diff_conflict_node.get("diff_branch_action"))),
-            diff_branch_value=diff_branch_value,
-            diff_branch_changed_at=Timestamp(diff_timestamp_str) if diff_timestamp_str else None,
-            selected_branch=ConflictSelection(selected_branch) if selected_branch else None,
-        )
-
-        linked_node.conflict = conflict
-        return conflict
