+++ backend/infrahub/dependencies/builder/diff/deserializer.py
@@ -1,0 +1,8 @@
+from infrahub.core.diff.repository.deserializer import EnrichedDiffDeserializer
+from infrahub.dependencies.interface import DependencyBuilder, DependencyBuilderContext
+
+
+class DiffDeserializerDependency(DependencyBuilder[EnrichedDiffDeserializer]):
+    @classmethod
+    def build(cls, context: DependencyBuilderContext) -> EnrichedDiffDeserializer:
+        return EnrichedDiffDeserializer()
